Section1: Game Setup
Created a blank 3rd person character setup with all assets imported that I will be using for this game. The next commit will jump right into developing the character, complete with movement and basic attacks.
As a side note, this is the first game that I will be coding with Rider. In most cases, live coding will not be possible without having to constantly shut the engine down when creating C++ scripts, which was required with Visual Studios.