Section 5: Enemy AI
Created an Enemy behavior tree that allows enemies to perceive (see) the Player character and attack when appropriate. Implemented a crowd following component so Enemies avoid each other when trying to get to the Player. Also implemented an Environment Query System so enemies can select appropriate places to move to when strafing. Added appropriate montages for walking, strafing, and attacking. Also enabled motion warping when attacking, so the Enemy is always facing the Player during the attack. Having some logic issues with the BT to abort logic when a Character is dead, but hoping to get that resolved before the next commit. Next commit will generally focus on Hero combat abilities for rolling, hit reaction, blocking, and enemy target lock. A lot of work, so will be using a different branch for these changes. Then I'll merge to the main branch when all combat abilities are implemented.