Section 6.2: Hero Combat Ability Hit React
Created a gameplay ability for hero hit reactions, which calculates the direction from which the hero character is being hit by the enemy. Created montages for front, back, left, and right hit reactions.  I also created a material instance that flashes when the hero character is hit. Finally, I added sound to all the hit react montages for quality immersion by the player.